finalProject/backend/requirements.txt
# backend/requirements.txt

fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
gcsfs
tensorflow
matplotlib
pandas
keras